chore: 멤버 롤 변경 UI dialog로 변경(#533)
