+++ apps/admin/src/widgets/bottom-sheet/model/constants.ts
@@ -1,4 +1,3 @@
-import { ChangeRoleSheet } from '@/features/member/role-change/ui/ChangeRoleSheet';
 import { SignupRequestBottomSheet } from '@/features/signup-request/ui/SignupRequestBottomSheet';
 import type { BottomSheetMap, BottomSheetType } from '@/shared/store/bottomSheetStore';
 import { MemberManagementSheet } from '@/widgets/member-management-sheet/MemberManagementSheet';
@@ -16,5 +15,4 @@ type BottomSheetComponents = {
 export const SHEET_COMPONENTS = {
   signup: SignupRequestBottomSheet,
   member: MemberManagementSheet,
-  changeRole: ChangeRoleSheet,
 } satisfies BottomSheetComponents;
